UI 1차 완료

+++ Assets/Feature_Player/script/Player/PlayerResource.cs
@@ -9,7 +9,7 @@ using static UnityEngine.Rendering.DebugUI;
 
 public class PlayerResource : ScriptableObject
 {
-    //플레이어 체력
+    //플레이어 최대체력
     public int maxHP = 50;
 
     //플레이어 이동속도
@@ -30,22 +30,22 @@ public class PlayerResource : ScriptableObject
         maxHP += value;
         return maxHP;
     }
-    public float UpSpeed(int value)
+    public float UpSpeed(float value)
     {
         speed += value;
         return speed;
     }
-    public float UpJumpForce(int value)
+    public float UpJumpForce(float value)
     {
         jumpForce += value;
         return jumpForce;
     }
-    public float UpAttackSpeed(int value)
+    public float UpAttackSpeed(float value)
     {
         attackSpeed += value;
         return attackSpeed;
     }
-    public float UpAttackDamage(int value)
+    public float UpAttackDamage(float value)
     {
         attackDamage += value;
         return attackDamage;
